+++ libc/test/src/stdlib/strtoul_test.cpp
@@ -147,7 +147,7 @@ static char int_to_b36_char(int input) {
 
 TEST(LlvmLibcStrToULTest, DecodeInOtherBases) {
   char small_string[4] = {'\0', '\0', '\0', '\0'};
-  for (unsigned int base = 2; base <= 36; ++base) {
+  for (unsigned long base = 2; base <= 36; ++base) {
     for (unsigned long first_digit = 0; first_digit <= 36; ++first_digit) {
       small_string[0] = int_to_b36_char(first_digit);
       if (first_digit < base) {
@@ -163,7 +163,7 @@ TEST(LlvmLibcStrToULTest, DecodeInOtherBases) {
     }
   }
 
-  for (unsigned int base = 2; base <= 36; ++base) {
+  for (unsigned long base = 2; base <= 36; ++base) {
     for (unsigned long first_digit = 0; first_digit <= 36; ++first_digit) {
       small_string[0] = int_to_b36_char(first_digit);
       for (unsigned long second_digit = 0; second_digit <= 36; ++second_digit) {
@@ -187,7 +187,7 @@ TEST(LlvmLibcStrToULTest, DecodeInOtherBases) {
     }
   }
 
-  for (unsigned int base = 2; base <= 36; ++base) {
+  for (unsigned long base = 2; base <= 36; ++base) {
     for (unsigned long first_digit = 0; first_digit <= 36; ++first_digit) {
       small_string[0] = int_to_b36_char(first_digit);
       for (unsigned long second_digit = 0; second_digit <= 36; ++second_digit) {
